refactor: break internal cmd helper up over multiple files
